requirements.txt
numpy
matplotlib
hydra-core